tree truncs
