+++ ui/src/app/cohort-search/modal/modal.component.css
@@ -91,9 +91,50 @@ button.active {
 }
 
 .panel.show {
-  display: block;
+  display: flex;
   animation: fadeEffect 1s;
 }
+
+.panel-left {
+  flex: 1;
+  min-width: 14rem;
+  overflow-y: auto;
+  overflow-x: hidden;
+  width: 100%;
+  height: 100%;
+  padding-right: 0.4rem;
+}
+
+.panel-right.selected {
+  width: 14rem;
+  min-width: 14rem;
+  border: 2px solid #d7d7d7;
+  height: 100%;
+  border-radius: 5px;
+  overflow-y: auto;
+  overflow-x: hidden;
+}
+
+h5.selection-title {
+  margin-top: 0;
+}
+
+@media all and (max-width: 900px) {
+  .panel {
+    flex-flow: column nowrap;
+    justify-content: space-between;
+  }
+  .panel-right.selected {
+    flex: 1;
+    width: 100%;
+    min-height: 50%;
+  }
+  .panel-left {
+    min-height: 50%;
+    padding-bottom: 0.2rem;
+  }
+}
+
 
 @keyframes fadeEffect {
   from {opacity: 0;}
