+++ ui/src/app/cohort-search/node-info/node-info.component.html
@@ -33,4 +33,3 @@
     </span>
   </div>
 </div>
-
