--- ui/src/app/cohort-search/tree-page/tree-page.component.css
@@ -1,58 +1,0 @@
-.page {
-  height: 100%;
-  display: flex;
-}
-
-.tree-view {
-  min-width: 14rem;
-  flex: 1;
-}
-
-.node-list {
-  overflow-y: auto;
-  overflow-x: hidden;
-  width: 100%;
-  height: 100%;
-  padding-right: 0.4rem;
-}
-
-.selected {
-  width: 14rem;
-  min-width: 14rem;
-  border: 2px solid #d7d7d7;
-  height: 100%;
-  padding-left: 0.4rem;
-  border-radius: 5px;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.selection-title {
-  margin-top: 0;
-}
-
-@media all and (max-width: 900px) {
-  .page {
-    flex-flow: column nowrap;
-    justify-content: space-between;
-  }
-  .selected {
-    flex: 1;
-    width: 100%;
-    min-height: 50%;
-  }
-  .tree-view {
-    min-height: 50%;
-    padding-bottom: 0.2rem;
-  }
-  .node-list {
-    padding-right: 0;
-  }
-}
-
-/**
- * We have to override this so that the text will truncate correctly
- */
-::ng-deep .clr-treenode-content {
-  width: 100%!important;
-}
