--- ui/src/app/cohort-search/tree-page/tree-page.component.html
@@ -1,25 +1,0 @@
-<div class="page">
-  <div class="tree-view">
-
-    <div *ngIf="loading" class="spinner" [style.left.%]="50">
-      Loading...
-    </div>
-
-    <div *ngIf="error">
-      <em>Error! Oh Noes!</em>
-    </div>
-
-    <div class="node-list">
-      <crit-node *ngFor="let item of children" [node]="item"></crit-node>
-    </div>
-  </div>
-
-  <div class="selected">
-    <h5 class="selection-title">{{ selectionTitle }}</h5>
-    <crit-selection-info
-      *ngFor="let parameter of (criteriaList$ | async); let i = index"
-      [index]="i"
-      [parameter]="parameter">
-    </crit-selection-info>
-  </div>
-</div>
