+++ ui/src/app/cohort-search/tree/tree.component.css
@@ -1,0 +1,6 @@
+/**
+ * We have to override this so that the text will truncate correctly
+ */
+::ng-deep .clr-treenode-content {
+  width: 100%!important;
+}
